Put `LabeledItems` and friends into a separate module.
