+++ dex.cabal
@@ -49,7 +49,7 @@ library
                        Builder, Cat, Export,
                        Simplify, TopLevel,
                        Autodiff, Interpreter, Logging, CUDA,
-                       LLVM.JIT, LLVM.Shims, Err
+                       LLVM.JIT, LLVM.Shims, Err, LabeledItems
   if flag(live)
     exposed-modules:   Actor, RenderHtml, LiveOutput
   other-modules:       Paths_dex
